requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
base58==2.1.1
bitarray==2.6.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
cycler==0.11.0
cytoolz==0.12.0
Deprecated==1.2.13
duneapi==4.0.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
fonttools==4.37.1
frozenlist==1.3.1
hexbytes==0.3.0
idna==3.3
ipfshttpclient==0.8.0a2
joblib==1.1.0
jsonschema==4.16.0
kiwisolver==1.4.4
lru-dict==1.1.8
matplotlib==3.5.3
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
numpy==1.23.3
packaging==21.3
pandas==1.4.4
parsimonious==0.8.1
Pillow==9.2.0
protobuf==3.20.1
pycodestyle==2.9.1
pycryptodome==3.15.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
rlp==2.0.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
types-Deprecated==1.2.9
types-requests==2.28.10
types-urllib3==1.26.24
urllib3==1.26.12
varint==1.0.2
web3==5.30.0
websockets==9.1
wrapt==1.14.1
yarl==1.8.1
